midi入力による音の強弱実装
