+++ src/main/java/com/github/te170476/musicdsl/midi/MidiInputReceiver.java
@@ -6,9 +6,10 @@ import javax.sound.midi.ShortMessage;
 import java.util.Set;
 
 public class MidiInputReceiver implements Receiver {
+
     public final String name;
-    private final Set<Integer> onPressKeySet;
-    public MidiInputReceiver(String name, Set<Integer> onPressKeySet){
+    private final Set<MidiInput> onPressKeySet;
+    public MidiInputReceiver(String name, Set<MidiInput> onPressKeySet){
         this.name = name;
         this.onPressKeySet = onPressKeySet;
         System.out.println("midi connect");
@@ -20,26 +21,29 @@ public class MidiInputReceiver implements Receiver {
 
         ShortMessage shortMessage = ((ShortMessage)message);
         System.out.println(shortMessage.getCommand());
-        System.out.println(shortMessage.getData1());
-        System.out.println(shortMessage.getData2());
-        var keyNumber = shortMessage.getData1();
-        switch (shortMessage.getCommand()) {
+        var command = shortMessage.getCommand();
+        var keyCode = shortMessage.getData1();
+        var strength = shortMessage.getData2();
+        System.out.println(keyCode);
+        System.out.println(strength);
+        var input = new MidiInput(keyCode, strength);
+        switch (command) {
             case 144:
-                onPress(keyNumber);
+                onPress(input);
                 break;
             case 128:
-                onRelease(keyNumber);
+                onRelease(input);
                 break;
             default:
                 break;
         }
         System.out.println("midi received");
     }
-    public void onPress(int key) {
-        onPressKeySet.add(key);
+    public void onPress(MidiInput input) {
+        onPressKeySet.add(input);
     }
-    public void onRelease(int key) {
-        onPressKeySet.remove(key);
+    public void onRelease(MidiInput input) {
+        onPressKeySet.remove(input);
     }
 
     @Override
